display participants grid on keyword page
