+++ src/components/Participants/Participants.vue
@@ -1,84 +1,19 @@
 <template>
   <div class="participants">
     <h1 class="pageTitle">Participantes</h1>
-    <Loader :loading="loadingData">
-      <div class="participants">
-        <div
-          class="participant"
-          v-for="person in participants"
-          v-bind:key="person.name"
-          v-bind:style="{
-            'background-image': person.img ? `url(${person.img})` : 'none'
-          }"
-        >
-          <h3 class="participantName">
-            <router-link :title="person.name" :to="urls.participant(person.slug)">
-              {{ person.name }}
-            </router-link>
-          </h3>
-        </div>
-      </div>
-    </Loader>
+    <ParticipantsGrid />
   </div>
 </template>
 
 <script>
 import { Vue, Component } from 'vue-property-decorator'
-import { loadParticipants } from '../../../api/client'
-import { getKeywordsFromParticipants } from '@/helpers/apiHelpers'
-import store from '@/config/store'
-import Loader from '@/components/common/Loader'
+import ParticipantsGrid from '@/components/Participants/ParticipantsGrid'
 
 @Component({
-  store,
   components: {
-    Loader
+    ParticipantsGrid
   }
 })
-class Participants extends Vue {
-  urls = this.$root.urls
-
-  loadingData = false
-
-  get participants () {
-    return this.$store.state.participants
-  }
-
-  mounted () {
-    this.loadingData = true
-    loadParticipants().then((response) => {
-      this.$store.commit('loadParticipants', response)
-      this.$store.commit('loadKeywords', getKeywordsFromParticipants(response))
-      this.loadingData = false
-    })
-  }
-}
+class Participants extends Vue {}
 export default Participants
 </script>
-
-<style scoped lang="stylus">
-@import '../../styles/colors'
-
-.participants
-  margin-top 10px
-  display flex
-  flex-wrap wrap
-
-  .participant
-    min-height 150px
-    padding 10px
-    display flex
-    align-items center
-    flex 1 1 20%
-    max-width 20%
-    margin-bottom 10px
-    background-size  0 0
-    background-repeat no-repeat
-    background-position center center
-    transition .2s all
-    &:hover
-      background-size cover
-      background-color darken($purple, 30%)
-    .participantName
-      opacity 1
-</style>
